+++ app/controllers/work_packages_controller.rb
@@ -377,15 +377,15 @@ class WorkPackagesController < ApplicationController
   end
 
   def relations
-    @relations ||= work_package.relations.includes(:issue_from => [:status,
+    @relations ||= work_package.relations.includes(:from => [:status,
                                                                    :priority,
                                                                    :type,
                                                                    { :project => :enabled_modules }],
-                                                   :issue_to => [:status,
+                                                   :to => [:status,
                                                                  :priority,
                                                                  :type,
                                                                  { :project => :enabled_modules }])
-                                         .select{ |r| r.other_issue(work_package) && r.other_issue(work_package).visible? }
+                                         .select{ |r| r.other_work_package(work_package) && r.other_work_package(work_package).visible? }
   end
 
   def priorities
@@ -426,7 +426,7 @@ class WorkPackagesController < ApplicationController
   def per_page_param
     case params[:format]
     when 'csv', 'pdf'
-      Setting.issues_export_limit.to_i
+      Setting.work_packages_export_limit.to_i
     when 'atom'
       Setting.feeds_limit.to_i
     else