+++ app/models/project.rb
@@ -45,11 +45,8 @@ class Project < ActiveRecord::Base
   has_many :members, :include => [:user, :roles], :conditions => "#{User.table_name}.type='User' AND #{User.table_name}.status=#{User::STATUSES[:active]}"
   has_many :assignable_members,
            :class_name => 'Member',
-           :include => [:user, :roles],
-           :conditions => ["#{User.table_name}.type=? AND #{User.table_name}.status=? AND roles.assignable = ?",
-                           'User',
-                           User::STATUSES[:active],
-                           true]
+           :include => [:principal, :roles],
+           :conditions => Proc.new { self.class.assignable_members_condition }
   has_many :memberships, :class_name => 'Member'
   has_many :member_principals, :class_name => 'Member',
                                :include => :principal,
@@ -596,9 +593,9 @@ class Project < ActiveRecord::Base
     Member.delete_all(['project_id = ?', id])
   end
 
-  # Users issues can be assigned to
+  # Users/groups a work_package can be assigned to
   def assignable_users
-    assignable_members.map(&:user).sort
+    assignable_members.map(&:principal).compact.sort
   end
 
   # Returns the mail adresses of users that should be always notified on project events
@@ -990,23 +987,23 @@ class Project < ActiveRecord::Base
 
       # Relations
       issue.relations_from.each do |source_relation|
-        new_issue_relation = IssueRelation.new
-        new_issue_relation.force_attributes = source_relation.attributes.dup.except("id", "work_package_from_id", "work_package_to_id")
-        new_issue_relation.issue_to = work_packages_map[source_relation.issue_to_id]
-        if new_issue_relation.issue_to.nil? && Setting.cross_project_issue_relations?
-          new_issue_relation.issue_to = source_relation.issue_to
+        new_relation = Relation.new
+        new_relation.force_attributes = source_relation.attributes.dup.except("id", "work_package_from_id", "work_package_to_id")
+        new_relation.to = work_packages_map[source_relation.to_id]
+        if new_relation.to.nil? && Setting.cross_project_work_package_relations?
+          new_relation.to = source_relation.to
         end
-        new_issue.relations_from << new_issue_relation
+        new_issue.relations_from << new_relation
       end
 
       issue.relations_to.each do |source_relation|
-        new_issue_relation = IssueRelation.new
-        new_issue_relation.force_attributes = source_relation.attributes.dup.except("id", "work_package_from_id", "work_package_to_id")
-        new_issue_relation.issue_from = work_packages_map[source_relation.issue_from_id]
-        if new_issue_relation.issue_from.nil? && Setting.cross_project_issue_relations?
-          new_issue_relation.issue_from = source_relation.issue_from
+        new_relation = Relation.new
+        new_relation.force_attributes = source_relation.attributes.dup.except("id", "work_package_from_id", "work_package_to_id")
+        new_relation.from = work_packages_map[source_relation.from_id]
+        if new_relation.from.nil? && Setting.cross_project_work_package_relations?
+          new_relation.from = source_relation.from
         end
-        new_issue.relations_to << new_issue_relation
+        new_issue.relations_to << new_relation
       end
     end
   end
@@ -1108,5 +1105,20 @@ class Project < ActiveRecord::Base
       subproject.send :archive!
     end
     update_attribute :status, STATUS_ARCHIVED
+  end
+
+  protected
+
+  def self.assignable_members_condition
+
+    condition = Setting.work_package_group_assignment? ?
+                  ["(#{Principal.table_name}.type=? OR #{Principal.table_name}.type=?)", 'User', 'Group'] :
+                  ["(#{Principal.table_name}.type=?)", 'User']
+
+    condition[0] += " AND #{User.table_name}.status=? AND roles.assignable = ?"
+    condition << User::STATUSES[:active]
+    condition << true
+
+    sanitize_sql_array condition
   end
 end