+++ app/models/query.rb
@@ -144,15 +144,14 @@ class Query < ActiveRecord::Base
                            "estimated_hours" => { :type => :integer, :order => 13 },
                            "done_ratio" =>  { :type => :integer, :order => 14 }}
 
-    user_values = []
-    user_values << ["<< #{l(:label_me)} >>", "me"] if User.current.logged?
+    principals = []
     if project
-      user_values += project.users.sort.collect{|s| [s.name, s.id.to_s] }
+      principals += project.principals.sort
     else
       all_projects = Project.visible.all
       if all_projects.any?
         # members of visible projects
-        user_values += User.active.find(:all, :conditions => ["#{User.table_name}.id IN (SELECT DISTINCT user_id FROM members WHERE project_id IN (?))", all_projects.collect(&:id)]).sort.collect{|s| [s.name, s.id.to_s] }
+        principals += Principal.active.find(:all, :conditions => ["#{User.table_name}.id IN (SELECT DISTINCT user_id FROM members WHERE project_id IN (?))", all_projects.collect(&:id)]).sort
 
         # project filter
         project_values = []
@@ -163,8 +162,24 @@ class Query < ActiveRecord::Base
         @available_filters["project_id"] = { :type => :list, :order => 1, :values => project_values} unless project_values.empty?
       end
     end
-    @available_filters["assigned_to_id"] = { :type => :list_optional, :order => 4, :values => user_values } unless user_values.empty?
-    @available_filters["author_id"] = { :type => :list, :order => 5, :values => user_values } unless user_values.empty?
+    principals_by_class = principals.group_by(&:class)
+
+    user_values = principals_by_class[User].present? ?
+                    principals_by_class[User].collect{ |s| [s.name, s.id.to_s] }.sort :
+                    []
+
+    group_values = Setting.work_package_group_assignment? && principals_by_class[Group].present? ?
+                      principals_by_class[Group].collect{ |s| [s.name, s.id.to_s] }.sort :
+                      []
+
+    assigned_to_values = (user_values + group_values).sort
+    assigned_to_values = [["<< #{l(:label_me)} >>", "me"]] + assigned_to_values if User.current.logged?
+    @available_filters["assigned_to_id"] = { :type => :list_optional, :order => 4, :values => assigned_to_values } unless assigned_to_values.empty?
+
+    author_values = []
+    author_values << ["<< #{l(:label_me)} >>", "me"] if User.current.logged?
+    author_values += user_values
+    @available_filters["author_id"] = { :type => :list, :order => 5, :values => author_values } unless author_values.empty?
 
     group_values = Group.all.collect {|g| [g.name, g.id.to_s] }
     @available_filters["member_of_group"] = { :type => :list_optional, :order => 6, :values => group_values, :name => I18n.t('query_fields.member_of_group') } unless group_values.empty?
@@ -175,7 +190,8 @@ class Query < ActiveRecord::Base
     if User.current.logged?
       # populate the watcher list with the same user list as other user filters if the user has the :view_work_package_watchers permission in at least one project
       # TODO: this could be differentiated more, e.g. all users could watch issues in public projects, but won't necessarily be shown here
-      watcher_values = User.current.allowed_to_globally?(:view_work_package_watchers, {}) ? user_values : [["<< #{l(:label_me)} >>", "me"]]
+      watcher_values = [["<< #{l(:label_me)} >>", "me"]]
+      user_values.each { |v| watcher_values << v } if User.current.allowed_to_globally?(:view_work_packages_watchers, {})
       @available_filters["watcher_id"] = { :type => :list, :order => 15, :values => watcher_values }
     end
 
@@ -292,7 +308,7 @@ class Query < ActiveRecord::Base
     if has_default_columns?
       available_columns.select do |c|
         # Adds the project column by default for cross-project lists
-        Setting.issue_list_default_columns.include?(c.name.to_s) || (c.name == :project && project.nil?)
+        Setting.work_package_list_default_columns.include?(c.name.to_s) || (c.name == :project && project.nil?)
       end
     else
       # preserve the column_names order
@@ -306,7 +322,7 @@ class Query < ActiveRecord::Base
       names = names.select {|n| n.is_a?(Symbol) || !n.blank? }
       names = names.collect {|n| n.is_a?(Symbol) ? n : n.to_sym }
       # Set column_names to nil if default columns
-      if names.map(&:to_s) == Setting.issue_list_default_columns
+      if names.map(&:to_s) == Setting.work_package_list_default_columns
         names = nil
       end
     end
@@ -361,7 +377,7 @@ class Query < ActiveRecord::Base
   end
 
   def any_summable_columns?
-    Setting.issue_list_summable_columns.any?
+    Setting.work_package_list_summable_columns.any?
   end
 
   def group_by_column
@@ -387,7 +403,7 @@ class Query < ActiveRecord::Base
           # all subprojects
           ids += project.descendants.collect(&:id)
         end
-      elsif Setting.display_subprojects_issues?
+      elsif Setting.display_subprojects_work_packages?
         ids += project.descendants.collect(&:id)
       end
       project_clauses << "#{Project.table_name}.id IN (%s)" % ids.join(',')
@@ -409,7 +425,14 @@ class Query < ActiveRecord::Base
 
       # "me" value subsitution
       if %w(assigned_to_id author_id watcher_id).include?(field)
-        v.push(User.current.logged? ? User.current.id.to_s : "0") if v.delete("me")
+        if v.delete("me")
+          if User.current.logged?
+            v.push(User.current.id.to_s)
+            v += User.current.group_ids.map(&:to_s) if field == 'assigned_to_id'
+          else
+            v.push("0")
+          end
+        end
       end
 
       sql = ''