+++ spec/models/work_package_spec.rb
@@ -95,6 +95,21 @@ describe WorkPackage do
         it { should be_nil }
       end
     end
+
+    describe :assigned_to do
+      context :group_assignment do
+        let(:group) { FactoryGirl.create(:group) }
+
+        before do
+          Setting.stub(:work_package_group_assignment).and_return(true)
+        end
+
+        subject { FactoryGirl.create(:work_package,
+                                     assigned_to: group).assigned_to }
+
+        it { should eq(group) }
+      end
+    end
   end
 
   describe :type do
@@ -163,6 +178,32 @@ describe WorkPackage do
       it 'should return all users the project deems to be assignable' do
         should include(user)
       end
+    end
+
+    context "with work_package_group_assignment" do
+      let(:group) { FactoryGirl.create(:group) }
+      let(:work_package) { FactoryGirl.create(:work_package) }
+
+      before do
+        Setting.stub(:work_package_group_assignment?).and_return(true)
+        work_package.project.add_member! group, FactoryGirl.create(:role)
+      end
+
+      subject { work_package.assignable_users }
+      it { should include(group) }
+    end
+
+    context "without work_package_group_assignment" do
+      let(:group) { FactoryGirl.create(:group) }
+      let(:work_package) { FactoryGirl.create(:work_package) }
+
+      before do
+        Setting.stub(:work_package_group_assignment?).and_return(false)
+        work_package.project.add_member! group, FactoryGirl.create(:role)
+      end
+
+      subject { work_package.assignable_users }
+      it { should_not include(group) }
     end
 
     context "multiple users" do
@@ -302,7 +343,7 @@ describe WorkPackage do
           end
 
           shared_examples_for "save with open version" do
-            before do 
+            before do
               work_package.status = status_open
               work_package.fixed_version = version_open
             end
@@ -315,7 +356,7 @@ describe WorkPackage do
           context "in closed version" do
             include_context "in closed version"
 
-            before do 
+            before do
               work_package.status = status_open
               work_package.save
             end
@@ -373,13 +414,13 @@ describe WorkPackage do
       end
 
       context "time entry 1" do
-        subject { work_package.time_entries } 
+        subject { work_package.time_entries }
 
         it { should include(time_entry_1) }
       end
 
       context "time entry 2" do
-        subject { work_package.time_entries } 
+        subject { work_package.time_entries }
 
         it { should include(time_entry_2) }
       end
@@ -414,7 +455,7 @@ describe WorkPackage do
 
           it { should eq(target_category.id) }
         end
-        
+
         it_behaves_like "moved work package"
       end
 
@@ -540,7 +581,7 @@ describe WorkPackage do
 
     describe :value do
       context "work package field" do
-        before { Setting.stub(:issue_done_ratio).and_return 'issue_field' }
+        before { Setting.stub(:work_package_done_ratio).and_return 'issue_field' }
 
         context "work package 1" do
           subject { work_package_1.done_ratio }
@@ -556,7 +597,7 @@ describe WorkPackage do
       end
 
       context "work package status" do
-        before { Setting.stub(:issue_done_ratio).and_return 'issue_status' }
+        before { Setting.stub(:work_package_done_ratio).and_return 'issue_status' }
 
         context "work package 1" do
           subject { work_package_1.done_ratio }
@@ -575,7 +616,7 @@ describe WorkPackage do
     describe :update_done_ratio_from_issue_status do
       context "work package field" do
         before do
-          Setting.stub(:issue_done_ratio).and_return 'issue_field'
+          Setting.stub(:work_package_done_ratio).and_return 'issue_field'
 
           work_package_1.update_done_ratio_from_issue_status
           work_package_2.update_done_ratio_from_issue_status
@@ -589,7 +630,7 @@ describe WorkPackage do
 
       context "work package status" do
         before do
-          Setting.stub(:issue_done_ratio).and_return 'issue_status'
+          Setting.stub(:work_package_done_ratio).and_return 'issue_status'
 
           work_package_1.update_done_ratio_from_issue_status
           work_package_2.update_done_ratio_from_issue_status
@@ -1257,7 +1298,7 @@ describe WorkPackage do
       let(:closed_issue_status) { FactoryGirl.create(:closed_issue_status) }
 
       before do
-        Setting.stub(:issue_done_ratio).and_return('disabled')
+        Setting.stub(:work_package_done_ratio).and_return('disabled')
       end
 
       it 'should not update the work package done_ratio' do