+++ doc/CHANGELOG.md
@@ -29,13 +29,27 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
+* `#1715` Group assigned work packages
+* `#1770` New Comment Section layout errors
+* `#1790` Fix activity view bug coming up during the meeting adaptions to acts_as_journalized
 * `#1793` Data Migration Journals
+* `#1977` Set default type for planning elements
+* `#1990` Migrate issue relation
+* `#1997` Migrate journal activities
+* `#2008` Migrate attachments
+
+## 3.0.0pre17
+
+* `#1323` Wrong Calendarweek in Datepicker, replaced built in datepicker with jQuery UI datepicker
+* `#1843` Editing Membership Duration in admin area fails
 * `#1913` [Timelines] Enable drag&drop for select2 items in order to rearrange the order of the columns
+* `#1934` [Timelines] Table Loading takes really long
 * `#1978` Migrate legacy issues
 * `#1979` Migrate legacy planning elements
 * `#1982` Migrate planning element types
 * `#1983` Migrate queries
 * `#1987` Migrate user rights
+* `#1988` Migrate settings
 * `#2019` Migrate auto completes controller tests
 * `#2078` Work package query produces 500 when grouping on exclusively empty values